\name{colorTable}


\alias{colorTable}


\title{Table of Colors}


\description{
    
    Displays a Table of color codes and plots the colors themselves.
        
}
    
    

\usage{ 
colorTable(cex = 0.7)
}


\arguments{
  
    \item{cex}{
        a numeric value, determines the character size in the
        color plot, the default size is 0.7.
        }  
    
}


\value{
    
    \code{colorTable}\cr
    \cr
    
    returns a table of plot plot colors with the associated color numbers.
    
}


\author{

    Diethelm Wuertz for the Rmetrics \R-port.
    
}


\examples{   
## Color Table:
   colorTable()
}


\keyword{programming}

