\name{listDescription}


\alias{listDescription}


\title{Description File Listing}


\description{

    Lists the content of a description file. 
    
}


\usage{
listDescription(package = "Rmetrics")
}


\arguments{

    \item{package}{
        a character string denoting the name of the package
        to be listed.
        }
        
}


\value{

    prints the description file.

}


\examples{
## listDescription -
   listDescription("fUtilities")
}


\keyword{programming}

