\name{listIndex}


\alias{listIndex}


\title{Index File Listing}


\description{

    Lists the content of an index file. 
    
}


\usage{
listIndex(package = "Rmetrics")
}


\arguments{

    \item{package}{
        a character string denoting the name of the package
        to be listed.
        }
        
}


\value{

    prints the index file.

}
    
    
\author{

    Diethelm Wuertz for this Rmetrics port.
    
}


\examples{
## listIndex -
   listIndex("fUtilities")
}


\keyword{programming}

